requirements.txt
psutil==5.9.8
PyMySQL==1.1.1
PyYAML==6.0.2
